Move UpdateStrategy classes to strategy package
item 패키지를 만들까 했는데, depth 증가보다는 strategy만 두는거로..
