+++ Java/src/main/java/com/gildedrose/strategy/AgedBrieUpdateStrategy.java
@@ -1,4 +1,6 @@
-package com.gildedrose;
+package com.gildedrose.strategy;
+
+import com.gildedrose.Item;
 
 public class AgedBrieUpdateStrategy implements ItemUpdateStrategy {
     @Override
