+++ Java/src/main/java/com/gildedrose/strategy/ItemUpdateStrategy.java
@@ -1,4 +1,6 @@
-package com.gildedrose;
+package com.gildedrose.strategy;
+
+import com.gildedrose.Item;
 
 public interface ItemUpdateStrategy {
     void update(Item item);
