+++ Java/src/main/java/com/gildedrose/strategy/StandardItemUpdateStrategy.java
@@ -1,4 +1,6 @@
-package com.gildedrose;
+package com.gildedrose.strategy;
+
+import com.gildedrose.Item;
 
 public class StandardItemUpdateStrategy implements ItemUpdateStrategy {
     @Override
