+++ Java/src/main/java/com/gildedrose/strategy/SulfurasUpdateStrategy.java
@@ -1,4 +1,6 @@
-package com.gildedrose;
+package com.gildedrose.strategy;
+
+import com.gildedrose.Item;
 
 public class SulfurasUpdateStrategy implements ItemUpdateStrategy {
     @Override
